Adding fading lightning bolts
The lightning bolts sort of fade after more lightning bolts are clicked
in. The background also now switches to black and displays "You must hate
this table. Shame on you,". Also the canvas can be reset with the letter
r with the keypressed function. I also added some comments in order to
separate and describe what those are. I don't think they're needed but
it's a good reminder. Made these changes as I wanted lightning to fade
and the lightning to be reset.